Remoe Joi
After profiling Joi was taking up 50% of the cpu usage to just parse the payload coming in :O why? lol. There is still other optimizations that need to be done but this is the most used function and slowest function in the flow so its a big win. Roughly 2X the performance now