customers_v2 elasticsearch index
Create the new v2 index. Drop beacon configuration visits as we are not entirely sure how we will use this right now. This can easily be added in later.

Changed versions from strings to objects. Allows for better aggregation and filtering

Remove ‘customers’ as a routing key for aliases. This routing key made sure all documents were routed to shard 0