Filter places by account
If two accounts have the same place registered postgres will select the first entry. This is in the case of the inbox app where the demo account had stephen’s house as well as the main account. Since we weren’t filtering by account id it was just selecting the first place that satisfied the conditions