Fix error handling while streaming back to the client (#148)
* Fix error handling while streaming back to the client

When a server is streaming back to a client it does not have the callback function. Instead it only operates on the writable stream. If we encounter an error then we just emit an `error` event on the writable stream. grpc then listens for this and terminates the call for us