[sdk-api] Auto set profile updated_at after update (#235)
We don’t get any information back from the audience service. To avoid version conflicts with whats up on elasticsearch now we set the current updated_at time to be an in memory date. This could be be behind a few seconds if network conditions are extrememly bad but the indexing is self healing as long as the device checks in and modifies any of its state