[proto/struct] do not distinguish between int and double (#849)
struct is meant to represent a json value. There is no distinction between double and integer values it is always up to the parser if they so wish to parse double as strings or not. This simplifies code as well since all integer values can be stored within a double value